General improvements like fixing warnings

+++ src/file_reader.c
@@ -20,7 +20,7 @@ char *read_file(char *path, char *fallback, char *read_error) {
       return read_error;
     }
     size_t bytes_read = fread(buffer, 1, file_size, response_fd);
-    if (bytes_read != file_size) {
+    if (bytes_read != (size_t) file_size) {
       perror("reading full file");
       free(buffer);
       fclose(response_fd);
